add csrf.token to all forms, fix Long Id usage

+++ src/main/java/org/JavaArt/TicketManager/DAO/ClientRepository.java
@@ -21,7 +21,7 @@ public interface ClientRepository {
 
     public void deleteClientsWithoutOrders(int minutes);
 
-    public Client getClientById(int id);
+    public Client getClientById(Long id);
 
     public List<Client> getAllClients();
